Partition test.js into another module, run_tests.js
run_tests.js, test.js
	-Isolate command-line argument handling code from that which
	runs test-suites and collects result statistics.
